v7.22 agent:v6.7

+++ AntDeployAgent/Model/Setting.cs
@@ -3,12 +3,13 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
 using System.Linq;
+using System.Threading.Tasks;
 
 namespace AntDeployAgentWindows.Model
 {
     public class Setting
     {
-        private static readonly System.Threading.Timer mDetectionTimer;
+        //private static readonly System.Threading.Timer mDetectionTimer;
         private static readonly int _clearOldPublishFolderOverDays = 10;
         private static readonly int _oldPulishLimit = 10;
         private static readonly List<string> MacWhiteList = new List<string>();
@@ -19,11 +20,11 @@ namespace AntDeployAgentWindows.Model
         public static bool NeedBackUp = true;
         static Setting()
         {
-#if DEBUG
-             mDetectionTimer = new System.Threading.Timer(OnVerify, null, 1000 * 5, 1000 * 5);
-#else
-             mDetectionTimer = new System.Threading.Timer(OnVerify, null, 1000 * 60 * 30, 1000 * 60 * 30);
-#endif
+            //#if DEBUG
+            //             mDetectionTimer = new System.Threading.Timer(OnVerify, null, 1000 * 5, 1000 * 5);
+            //#else
+            //             mDetectionTimer = new System.Threading.Timer(OnVerify, null, 1000 * 60 * 30, 1000 * 60 * 30);
+            //#endif
 
             var oldPulishLimit = System.Configuration.ConfigurationManager.AppSettings["OldPulishLimit"];
             if (int.TryParse(oldPulishLimit, out int value1) && value1 > 0)
@@ -32,7 +33,7 @@ namespace AntDeployAgentWindows.Model
             }
 
             var clearOldPublishFolderOverDaysStr = System.Configuration.ConfigurationManager.AppSettings["ClearOldPublishFolderOverDays"];
-            if (int.TryParse(clearOldPublishFolderOverDaysStr, out int value) && value>0)
+            if (int.TryParse(clearOldPublishFolderOverDaysStr, out int value) && value > 0)
             {
                 _clearOldPublishFolderOverDays = value;
             }
@@ -62,7 +63,7 @@ namespace AntDeployAgentWindows.Model
         public static string PublishWindowServicePathFolder = "";
         public static string BackUpWindowServicePathFolder = "";
 
-        public static void InitWebRoot(string rootPath,bool useCustomer = false)
+        public static void InitWebRoot(string rootPath, bool useCustomer = false)
         {
 
             if (string.IsNullOrEmpty(rootPath))
@@ -72,7 +73,7 @@ namespace AntDeployAgentWindows.Model
 
             WebRootPath = rootPath;
 
-            PublishPathFolder = (useCustomer)? rootPath: Path.Combine(WebRootPath, "antdeploy");
+            PublishPathFolder = (useCustomer) ? rootPath : Path.Combine(WebRootPath, "antdeploy");
 
             if (!Directory.Exists(PublishPathFolder))
             {
@@ -109,35 +110,55 @@ namespace AntDeployAgentWindows.Model
             }
         }
 
-
-        private static void OnVerify(object state)
+        public static void ClearOldFolders(bool isIis,string projectFolderName)
         {
-            mDetectionTimer.Change(-1, -1);
-            try
-            {
-                CheckOldFolder(PublishIIsPathFolder);
-                CheckOldFolder(BackUpIIsPathFolder);
-                CheckOldFolder(PublishWindowServicePathFolder);
-                CheckOldFolder(BackUpWindowServicePathFolder);
 
-            }
-            catch
+            new Task(() =>
             {
-                // ignored
-            }
-            finally
-            {
-#if DEBUG
-                mDetectionTimer.Change(1000 * 5, 1000 * 5);
-#else
-                mDetectionTimer.Change(1000 * 60 * 30, 1000 * 60 * 30);
-#endif
+                if (isIis)
+                {
+                    CheckOldFolder(PublishIIsPathFolder, projectFolderName);
+                    CheckOldFolder(BackUpIIsPathFolder, projectFolderName);
+                }
+                else
+                {
 
-            }
+                    CheckOldFolder(PublishWindowServicePathFolder, projectFolderName);
+                    CheckOldFolder(BackUpWindowServicePathFolder, projectFolderName);
+                }
+            
+            }).Start();
         }
 
+        //        private static void OnVerify(object state)
+        //        {
+        //            mDetectionTimer.Change(-1, -1);
+        //            try
+        //            {
+        //                ClearOldFolders();
+        //            }
+        //            catch
+        //            {
+        //                // ignored
+        //            }
+        //            finally
+        //            {
+        //#if DEBUG
+        //                mDetectionTimer.Change(1000 * 5, 1000 * 5);
+        //#else
+        //                mDetectionTimer.Change(1000 * 60 * 30, 1000 * 60 * 30);
+        //#endif
 
-        private static void CheckOldFolder(string path)
+        //            }
+        //        }
+
+
+        /// <summary>
+        /// 清除老的发布版本记录 (不会删除正在使用的版本 即使已经到期)
+        /// </summary>
+        /// <param name="path"></param>
+        /// <param name="projectFolder"></param>
+        private static void CheckOldFolder(string path, string projectFolder)
         {
             try
             {
@@ -146,20 +167,27 @@ namespace AntDeployAgentWindows.Model
 
                 if (!Directory.Exists(path)) return;
 
-                var applicationFolders = Directory.GetDirectories(path);
+                var applicationFolders = !string.IsNullOrEmpty(projectFolder) ? new List<string> { Path.Combine(path,projectFolder) }.ToArray() : Directory.GetDirectories(path);
                 if (applicationFolders.Length < 1) return;
 
                 foreach (var applicationFolder in applicationFolders)
                 {
                     var subFolders = Directory.GetDirectories(applicationFolder);
                     if (subFolders.Length < _oldPulishLimit) continue;//还没超过最低的保留记录数
+                    //找到current.txt文件 记录着当前正在使用的版本
+                    var currentText = Path.Combine(applicationFolder, "current.txt");
+                    var currentVersion = "";
+                    if (File.Exists(currentText))
+                    {
+                        currentVersion = File.ReadAllText(currentText);
+                    }
                     //超过了就要比对文件夹的日期 把超过了x天数的就要删掉
                     var oldFolderList = new List<OldFolder>();
                     foreach (var subFolder in subFolders)
                     {
                         var folder = new DirectoryInfo(subFolder);
                         var folderName = folder.Name;
-                        folderName = folderName.Replace("Backup", "").Replace("Err","").Replace("_", "");
+                        folderName = folderName.Replace("Backup", "").Replace("Err", "").Replace("_", "");
                         if (!DateTime.TryParseExact(folderName, "yyyyMMddHHmmss", null, DateTimeStyles.None, out DateTime createDate))
                         {
                             continue;
@@ -190,6 +218,11 @@ namespace AntDeployAgentWindows.Model
 
                     foreach (var target in targetList)
                     {
+                        //如果是当前正在使用的版本的话 不能删除
+                        if (!string.IsNullOrEmpty(currentVersion) && target.Name.Equals(currentVersion))
+                        {
+                            continue;
+                        }
                         try
                         {
                             Directory.Delete(target.FullName, true);
@@ -212,8 +245,8 @@ namespace AntDeployAgentWindows.Model
 
         public static void StopWatchFolderTask()
         {
-            mDetectionTimer.Change(-1, -1);
-            mDetectionTimer.Dispose();
+            //mDetectionTimer.Change(-1, -1);
+            //mDetectionTimer.Dispose();
         }
 
         /// <summary>
